requirements.txt
pandas
numpy
scikit-learn
tensorflow
opencv-python
pillow
matplotlib
fastapi
uvicorn
streamlit
langchain
langchain-openai
langchain-community
pinecone
pypdf
tiktoken
langchain-huggingface
rank_bm25
